+++ nightfall-deployer/contracts/Structures.sol
@@ -21,7 +21,7 @@ contract Structures {
 
     event CommittedToChallenge(bytes32 commitHash, address sender);
 
-    event InstantWithdrawalRequested(bytes32 withdrawTransactionHash, address paidBy, uint amount);
+    event InstantWithdrawalRequested(bytes32 withdrawTransactionHash, address paidBy, uint256 amount);
 
     /**
   These events are what the merkle-tree microservice's filters will listen for.
